+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 ---
 
+# [1.9.0] - 2021-01-04
+
+### Added
+
+- Pipiot - LevelSense
+- Pipiot - PeopleSense
+
 # [1.8.1] - 2022-01-03
 
 ### Changed